backend/requirements.txt
Django>=5.0
python-json-logger==2.0.7
psycopg[binary]
dj-database-url
python-dotenv
djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt
gunicorn
